requirements.txt
# المتطلبات الأساسية للمشروع

# Flask للـ API
Flask==3.0.0
Werkzeug==3.0.1

# python-telegram-bot v20+ (مع دعم async)
python-telegram-bot==20.7

# مكتبات إضافية
python-dotenv==1.0.0
requests==2.31.0

# قاعدة البيانات (اختياري)
# SQLAlchemy==2.0.23
# psycopg2-binary==2.9.9
